*** COMMITS INCLUDED IN gui_setup_advinst_opts__setup_part_2 ***
[SETUPLIB][BOOTDATA] Rename "MBRInstallType" to "BootLoaderLocation"

Less hardcoded references to "MBR" for code that should be generic
is always better.

[SETUPLIB] Move the initialization of other default values into InitializeSetup()

[SETUPLIB] Store a machine architecture type to determine e.g. how the bootloader has to be installed, etc.

Currently hardcoded for BIOS-based PC-AT; will be determined
at runtime in the future.

It will be used in later commits.

[SETUP:REACTOS] Redesign the "Advanced Installation Options" dialog.

In particular, use now a dynamic combo-box for listing the possible
locations where the bootloader can be installed.