*** COMMITS INCLUDED IN setup_partitioning_with_gui_setup_part_1 ***
[SETUPLIB] Add some device utility functions; to be used later.

[SETUPLIB][REACTOS][USETUP] Split FS-volume-specific functionality from partitions.

This greatly helps in reducing code complexity in some areas: code that
previously iterated over all partitions of a given disk, just to find
which ones were partitioned and contained a valid file system, now just
have to iterate over mounted volumes.
See in particular, lib/utils/osdetect.c and lib/fsutil.c .

- Remove FORMATSTATE "Preformatted" enum value;
- Cleanup osdetect code after introducing Volume support;
- Some simplifications for FormatState.

- Differentiate between 'new' partition and 'new' volume:

  * "New" partition: it has been created and added in the cached list,
    but not yet actually written into the disk.

  * "New" volume: newly-created volume (may be backed by a partition or
    not), not yet formatted. May exist on either new, or not new partition,
    or elsewhere.

[SETUPLIB] Don't immediately unmount volumes when deleting partitions...

...but wait until the partitions are actually updated on disk.

DeletePartition() deletes partitions only in the virtual partitions list
in memory; the partitions still exist on the disk and are written only
later when committing all partition changes.

Thus, we do the same for file system volumes that lay on top of these
partitions being deleted: we move the linked VOLENTRY structures to a
list of "pending volumes to unmount", while they still actually stay
mounted on the system. Only when partition changes are about to be
committed to disk, is the "pending volumes to unmount" list processed
and the volumes actually unmounted first.

[SETUPLIB] Support device naming for volumes

Split InitVolume() and setting a default volume device name.