** wip ** [NTOS:FSTUB] *ReadPartitionTable*(): PartitionNumber in the layout's PartitionEntry must be left uninitialized.
Addendum to commit e69f845dab.

This is a subtle observation.

First, it should be observed that when the partition manager driver
(or the disk class driver on Windows <= 2003) calls any of the
*ReadPartitionTable*() functions (and indirectly their helpers),
as part of IOCTL_DISK_GET_DRIVE_LAYOUT(_EX) handling, the retrieved
PartitionNumber values are reset and recalculated afterwards.
Meaning that whatever has been stored initially by the
*ReadPartitionTable*() functions is completely ignored.

Second, one should monitor (with a kernel debugger, e.g. WinDbg)
the ExAllocatePoolWithTag() calls done by the *ReadPartitionTable*()
functions themselves, when they (re-)allocate the partition buffer
that is going to be later returned to the caller. One then needs
to fill the allocated buffer (or calculate and fill only the
PartitionNumber of the layout entries) with a canary value.
Finally, once the *ReadPartitionTable*() functions return, one can
observe that the PartitionNumber have been left untouched.
On Windows, this means, left uninitialized (and filled with whatever
kernel data was present in memory from the ExAllocatePoolWithTag()
calls).

For ReactOS, we instead invalidate the PartitionNumber with -1,
a "known" value.